Numerov's method

Numerov's method

The Numerov method is introduced by Russian astronomer Boris Vasilyevich Numerov, to solve ODE

\[\frac{ d^2 y }{ dx^2 } + g(x) y(x) = s(x).\]

with initial conditions

\[y(x_0) = y_0 \quad y'(x_0) = y_0'.\]

Follow the derivation from the notes, we derive the $n + 1$ step of $y_{n + 1} = y(x_{n + 1})$ satisfies

\[\bigg( 1 + \frac{ h^2 }{ 12 } g_{n + 1} \bigg) y_{n + 1} = 2 y_{n} \bigg( 1 - \frac{ 5 h^2 }{ 12 } g_{n} \bigg) - \bigg( 1 + \frac{ h^2 }{ 12 } g_{n - 1} \bigg) y_{n - 1},\]

where $g_{n} = g(x_{n})$ and $s_{n} = s(x_{n})$, etc.

Therefore, we could split the $x$-axis to $N$ pieces, and solve the solution $y(x)$ piece-by-piece, i.e., from y[1] and y[2] to y[3], repeatedly.

If s(x) \equiv 0 on the domain of $x$, the the above equation simplifies to

\[y_{n + 1} = 2 \frac{ 12 - 5h^2 g_{n} }{ 12 + h^2 g_{n + 1} } y_{n} - y_{n - 1}.\]

NumericalMethodsInQuantumMechanics.EigenvalueProblems.NumerovMethod.numerov_iter — Method.

numerov_iter(y_prev, y, dx, gvec, svec)

Return the next step for the Numerov's method, given the current and previous step. See the detailed doc for this method.

Arguments

y_prev::Real: the y[n - 1] element.

y::Real: the y[n] element.

dx::Real: the step length, need to be small.

gvec::AbstractArray{<: Real}: stores g[n - 1], g[n] and g[n + 1].

svec::AbstractArray{<: Real}: stores s[n - 1], s[n] and s[n + 1].

See also: integrate

source

NumericalMethodsInQuantumMechanics.EigenvalueProblems.NumerovMethod.numerov_iter — Method.

numerov_iter(y_prev, y, dx, gvec)

Same as numerov_iter(y_prev, y, dx, gvec, svec), if $s(x) ≡ 0$ on the domain.

Arguments

y_prev::Real: the y[n - 1] element.

y::Real: the y[n] element.

dx::Real: the step length, need to be small.

gvec::AbstractArray{<: Real}: stores g[n - 1], g[n] and g[n + 1].

source

NumericalMethodsInQuantumMechanics.EigenvalueProblems.NumerovMethod.integrate — Method.

integrate(y0, yd0, r, gvec, svec)

Do the Numerov's method integration, return the solution $y(x)$, given that $g(x)$ and $s(x)$ already as vectors (already applied on $x$).

Arguments

y0::Real: the initial value $y(x_0) = y_0$, could be a guess.

yd0::Real: the initial value $y'(x_0) = y_0'$, could be a guess.

r::AbstractRange{<: Real}: a range, the domain $x$.

gvec::AbstractArray{<: Real}: the result of function $g$ applied on $x$ (range r).

svec::AbstractArray{<: Real}: the result of function $s$ applied on $x$ (range r).

source

NumericalMethodsInQuantumMechanics.EigenvalueProblems.NumerovMethod.integrate — Method.

integrate(y0, yd0, r, g, s)

Same as integrate(y0, yd0, r, gvec, svec), but g and s are two functions.

Arguments

y0::Real: the initial value $y(x_0) = y_0$, could be a guess.

yd0::Real: the initial value $y'(x_0) = y_0'$, could be a guess.

r::AbstractRange{<: Real}: a range, the domain $x$.

g::Function: the function $g$.

s::Function: the function $s$.

source