Numerov's method

The Numerov method is introduced by Russian astronomer Boris Vasilyevich Numerov, to solve ODE

\[\frac{ d^2 y }{ dx^2 } + g(x) y(x) = s(x).\]

with initial conditions

\[y(x_0) = y_0, \quad y'(x_0) = y_0'.\]

Follow the derivation from the notes, we derive the $n + 1$ step of $y_{n + 1} = y(x_{n + 1})$ satisfies

\[\bigg( 1 + \frac{ h^2 }{ 12 } g_{n + 1} \bigg) y_{n + 1} = 2 y_{n} \bigg( 1 - \frac{ 5 h^2 }{ 12 } g_{n} \bigg) - \bigg( 1 + \frac{ h^2 }{ 12 } g_{n - 1} \bigg) y_{n - 1},\]

where $g_{n} = g(x_{n})$ and $s_{n} = s(x_{n})$, etc.

Therefore, we could split the $x$-axis to $N$ pieces, and solve the solution $y(x)$ piece-by-piece, i.e., from y[1] and y[2] to y[3], repeatedly.

If $s(x) \equiv 0$ on the domain of $x$, the the above equation simplifies to

\[y_{n + 1} = 2 \frac{ 12 - 5h^2 g_{n} }{ 12 + h^2 g_{n + 1} } y_{n} - y_{n - 1}.\]

NumerovShooting.integrate — Function

integrate(ic, r, gvec, svec)

Do the Numerov's method integration, return the solution $y(x)$, given that $g(x)$ and $s(x)$ already as vectors (already applied on $x$).

Arguments

ic::InitialCondition: the initial condition $y(x_0) = y_0$ and $y'(x_0) = y_0'$, could be a guess.

r::AbstractRange{<:Real}: a range, the domain $x$.

gvec::AbstractArray{<:Real}: the result of function $g$ applied on $x$ (range r).

svec::AbstractArray{<:Real}: the result of function $s$ applied on $x$ (range r).